# internal/handler/user.go
package handler

import (
	"fmt"
	"log"
	"net/http"
	"time"

	"github.com/King0625/golang-todolist/internal/middleware"
	"github.com/King0625/golang-todolist/internal/model"
	"github.com/King0625/golang-todolist/internal/service"
	"github.com/King0625/golang-todolist/pkg/utils"
	"github.com/go-playground/validator/v10"
)

type RegisterPayload struct {
	Email     string `json:"email" validate:"required,email" example:"tdn@coat.jp"`
	FirstName string `json:"firstName" validate:"required,max=666" example:"yaju"`
	LastName  string `json:"lastName" validate:"required,max=666" example:"senpai"`
	Password  string `json:"password" validate:"required,min=6,max=12" example:"114514"`
}

type LoginPayload struct {
	Email    string `json:"email" validate:"required,email" example:"shxt@acceed.jp"`
	Password string `json:"password" validate:"required,min=6,max=12" example:"blackhole8"`
}

type LoginSuccessData struct {
	Token string `json:"token" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NTE5MDMzNTYsInVzZXJJRCI6MSwidXNlcm5hbWUiOiJLZW5DaGVuIn0.kBbaD4Z0Wacz-CBUacLq6fTpIdqq5Sfuoc4yF_BkfEE"`
}

type UserHandler struct {
	service  service.UserService
	validate *validator.Validate
}

func NewUserHandler(s service.UserService) *UserHandler {
	validate := validator.New()
	return &UserHandler{s, validate}
}

// Register godoc
// @Summary      Register a new user
// @Description  Register a user with email,firstName,lastName,password
// @Tags         users
// @Accept       json
// @Produce      json
// @Param        user  body  RegisterPayload  true  "user to create"
// @Success      201   {object}  utils.SuccessResponse
// @Failure      401   {object}  utils.ErrorResponse
// @Failure      400   {object}  utils.ErrorResponse
// @Router       /users/register [post]
func (h *UserHandler) Register(w http.ResponseWriter, r *http.Request) {
	var message string

	var payload RegisterPayload
	err := utils.ReadJSONRequest(w, r, &payload)
	if err != nil {
		log.Println(err)
		message = "cannot parse json body"
		utils.RespondError(w, http.StatusBadRequest, InvalidJSON, message, nil)
		return
	}

	if err = h.validate.Struct(payload); err != nil {
		message = "Validation failed"
		errs := err.(validator.ValidationErrors)
		details := make(map[string]string)
		for _, e := range errs {
			details[e.Field()] = e.ActualTag()
		}
		utils.RespondError(w, http.StatusBadRequest, ValidationError, message, details)
		return
	}

	currentTime := time.Now()

	// var uesr
	user := model.User{
		Email:     payload.Email,
		FirstName: payload.FirstName,
		LastName:  payload.LastName,
		Password:  payload.Password,
		CreatedAt: currentTime,
		UpdatedAt: currentTime,
	}

	err = h.service.Register(r.Context(), &user)
	if err != nil {
		message := "cannot insert user data into db"
		utils.RespondError(w, http.StatusInternalServerError, InternalError, message, nil)
		return
	}

	message = "register user successfully"
	utils.RespondSuccess(w, http.StatusCreated, message, nil)
}

// Login godoc
// @Summary      Login a user
// @Description  login with email,password
// @Tags         users
// @Accept       json
// @Produce      json
// @Param        user  body  LoginPayload  true  "user to login"
// @Success      200   {object}  utils.SuccessResponse{data=LoginSuccessData}
// @Failure      401   {object}  utils.ErrorResponse
// @Failure      400   {object}  utils.ErrorResponse
// @Router       /users/login [post]
func (h *UserHandler) Login(w http.ResponseWriter, r *http.Request) {
	var message string
	var payload LoginPayload
	err := utils.ReadJSONRequest(w, r, &payload)
	if err != nil {
		log.Fatal(err)
		message = "cannot parse json body"
		utils.RespondError(w, http.StatusBadRequest, InvalidJSON, message, nil)
		return
	}

	if err = h.validate.Struct(payload); err != nil {
		message = "validation failed"
		errs := err.(validator.ValidationErrors)
		details := make(map[string]string)
		for _, e := range errs {
			details[e.Field()] = e.ActualTag()
		}
		utils.RespondError(w, http.StatusBadRequest, ValidationError, message, details)
		return
	}

	user, err := h.service.Login(r.Context(), payload.Email, payload.Password)
	if err != nil {
		message = "login failed"
		utils.RespondError(w, http.StatusUnauthorized, Unauthorized, message, nil)
		return
	}

	jwtToken, err := utils.NewToken(user.FirstName+user.LastName, user.ID)
	if err != nil {
		message = "failed to issue jwt token from server"
		utils.RespondError(w, http.StatusInternalServerError, InternalError, message, nil)
		return
	}

	message = "login successfully"
	data := LoginSuccessData{jwtToken}
	utils.RespondSuccess(w, http.StatusOK, message, data)
}

// GetUserData godoc
// @Summary      get users data
// @Description  get user data
// @Tags         users
// @Accept       json
// @Security     BearerAuth
// @Produce      json
// @Success      200   {object}  utils.SuccessResponse{data=model.User}
// @Failure      401   {object}  utils.ErrorResponse
// @Failure      400   {object}  utils.ErrorResponse
// @Router       /users/me [get]
func (h *UserHandler) GetUserData(w http.ResponseWriter, r *http.Request) {
	var message string
	userID, ok := middleware.GetUserID(r)
	if !ok {
		message = "failed to fetch user identity from parsed jwt token"
		utils.RespondError(w, http.StatusUnauthorized, Unauthorized, message, nil)
		return
	}

	user, err := h.service.GetUserDataById(r.Context(), userID)
	if user == nil {
		fmt.Println(err)
		message = "user not found"
		utils.RespondError(w, http.StatusNotFound, UserNotFound, message, nil)
		return
	}

	message = "get user data successfully"
	utils.RespondSuccess(w, http.StatusOK, message, user)
}